pyproject.toml
[tool.poetry]
name = "pypulsar-rpc"
version = "0.1.0"
description = ""
authors = ["Kiryl Kovaliov <kkirusha@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pulsar-client = "2.10.1"
pydantic = "^1.9.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
poethepoet = "^0.16.0"
pylint = "^2.14.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
lint = "pylint --errors-only -v **/*.py"
